refactor(transforms): reorganize locations
