+++ packages/next-swc/crates/next-custom-transforms/src/transforms/amp_attributes.rs
