+++ packages/next-swc/crates/next-custom-transforms/src/transforms/cjs_optimizer.rs
