+++ packages/next-swc/crates/next-custom-transforms/src/transforms/import_analyzer.rs
