+++ packages/next-swc/crates/next-custom-transforms/src/transforms/next_ssg.rs
