+++ packages/next-swc/crates/next-custom-transforms/src/transforms/optimize_barrel.rs
