+++ packages/next-swc/crates/next-custom-transforms/src/transforms/optimize_server_react.rs
