+++ packages/next-swc/crates/next-custom-transforms/src/transforms/page_config.rs
