+++ packages/next-swc/crates/next-custom-transforms/src/transforms/shake_exports.rs
