+++ packages/next-swc/crates/wasm/src/lib.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use anyhow::{Context, Error};
 use js_sys::JsString;
-use next_custom_transforms::{custom_before_pass, TransformOptions};
+use next_custom_transforms::chain_transforms::{custom_before_pass, TransformOptions};
 use swc_core::common::Mark;
 use turbopack_binding::swc::core::{
     base::{
